+++ src/test/java/ladder/domain/person/PersonTest.java
@@ -1,6 +1,5 @@
-package ladder.domain;
+package ladder.domain.person;
 
-import ladder.domain.person.Person;
 import ladder.exception.person.PersonNameLengthException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
